Добавляет контроллер для отрисовки борда фильмов (без рефакторинга)
