+++ src/main.js
@@ -2,22 +2,23 @@ import HeaderProfileComponent from './components/header-profile';
 import MainMenuComponent from './components/main-menu';
 import StatisticComponent from './components/statistic';
 import SortComponent from './components/sort';
-import FilmCardComponent from './components/film-card';
-import ShowMoreBtnComponent from './components/show-more-btn';
+// import FilmCardComponent from './components/film-card';
+// import ShowMoreBtnComponent from './components/show-more-btn';
 import FilmListComponent from './components/film-list';
 import FilmCardContainerComponent from './components/film-card-container';
 import FooterStatisticComponent from './components/footer-statistic';
 
-import FilmDetailsPopupComponent from './components/film-details-popup';
+// import FilmDetailsPopupComponent from './components/film-details-popup';
 
 import {generateFilmCards} from './mock/film-card';
 
-import {render, RenderPosition, remove} from './utils/render';
-import NoFilmsComponent from './components/no-films';
+import {render, RenderPosition} from './utils/render';
+
+import PageController from './controllers/page-controller';
 
 const CARDS_COUNT = 10;
-const SHOWING_CARDS_COUNT_ON_START = 5;
-const SHOWING_CARDS_COUNT_BY_BUTTON = 5;
+// const SHOWING_CARDS_COUNT_ON_START = 5;
+// const SHOWING_CARDS_COUNT_BY_BUTTON = 5;
 
 const filmCards = generateFilmCards(CARDS_COUNT);
 
@@ -56,82 +57,88 @@ render(siteMain, new FilmListComponent(), RenderPosition.BEFOREEND);
 const siteFooter = document.querySelector(`.footer`);
 render(siteFooter, new FooterStatisticComponent(filmCards.length), RenderPosition.BEFOREEND);
 
+const filmsCardContanerComponent = new FilmCardContainerComponent();
 
-const renderFilmCard = (filmListElement, film) => {
-
-  const onFilmCardClick = () => {
-    render(siteFooter, filmDetailsPopupComponent, RenderPosition.AFTEREND);
-    document.addEventListener(`keydown`, onEscKeydown);
-  };
-
-  const onCloseButtonClick = () => {
-    document.querySelector(`body`).removeChild(filmDetailsPopupComponent.getElement());
-    document.removeEventListener(`keydown`, onEscKeydown);
-  };
-
-  const onEscKeydown = (evt) => {
-    const isEscape = evt.key === `Escape` || evt.key === `Esc`;
-
-    if (isEscape) {
-      document.querySelector(`body`).removeChild(filmDetailsPopupComponent.getElement());
-
-      document.removeEventListener(`keydown`, onEscKeydown);
-    }
-  };
-
-  const filmComponent = new FilmCardComponent(film);
-  filmComponent.setFilmCardClickHandler(onFilmCardClick);
-
-  const filmDetailsPopupComponent = new FilmDetailsPopupComponent(film);
-  filmDetailsPopupComponent.setPopupCloseButtonClickHandler(onCloseButtonClick);
-
-  render(filmListElement, filmComponent, RenderPosition.BEFOREEND);
-};
-
-const renderFilmsLists = (filmsComponent, films) => {
-
-  const filmsContainer = siteMain.querySelector(`.films`);
-
-  if (films.length === 0) {
-    filmsContainer.removeChild(filmsContainer.querySelector(`.films-list`));
-    render(filmsContainer, new NoFilmsComponent(), RenderPosition.BEFOREEND);
-    return;
-  }
-  render(filmsComponent.getElement(), new FilmListComponent(), RenderPosition.BEFOREEND);
-
-  const filmListElement = siteMain.querySelector(`.films-list`);
-  render(filmListElement, new FilmCardContainerComponent(), RenderPosition.BEFOREEND);
-  const mainFilmsListContainer = filmListElement.querySelector(`.films-list__container`);
-
-  let showingFilmsCount = SHOWING_CARDS_COUNT_ON_START;
-
-  films
-    .slice(0, showingFilmsCount)
-    .forEach((film) => {
-      renderFilmCard(mainFilmsListContainer, film);
-    });
+const pageController = new PageController(filmsCardContanerComponent);
+pageController.render(filmCards);
 
 
-  const showMoreButtonComponent = new ShowMoreBtnComponent();
+// const renderFilmCard = (filmListElement, film) => {
 
-  render(filmListElement, showMoreButtonComponent, `beforeend`);
+//   const onFilmCardClick = () => {
+//     render(siteFooter, filmDetailsPopupComponent, RenderPosition.AFTEREND);
+//     document.addEventListener(`keydown`, onEscKeydown);
+//   };
 
-  showMoreButtonComponent.setClickHandler(() => {
-    const prevFilmCount = showingFilmsCount;
-    showingFilmsCount = showingFilmsCount + SHOWING_CARDS_COUNT_BY_BUTTON;
+//   const onCloseButtonClick = () => {
+//     document.querySelector(`body`).removeChild(filmDetailsPopupComponent.getElement());
+//     document.removeEventListener(`keydown`, onEscKeydown);
+//   };
 
-    films.slice(prevFilmCount, showingFilmsCount)
-      .forEach((film) => renderFilmCard(mainFilmsListContainer, film));
-    if (showingFilmsCount >= films.length) {
-      remove(showMoreButtonComponent);
-    }
-  });
-};
+//   const onEscKeydown = (evt) => {
+//     const isEscape = evt.key === `Escape` || evt.key === `Esc`;
 
-const filmsComponent = new FilmCardContainerComponent();
+//     if (isEscape) {
+//       document.querySelector(`body`).removeChild(filmDetailsPopupComponent.getElement());
 
-renderFilmsLists(filmsComponent, filmCards);
+//       document.removeEventListener(`keydown`, onEscKeydown);
+//     }
+//   };
 
+//   const filmComponent = new FilmCardComponent(film);
+//   filmComponent.setFilmCardClickHandler(onFilmCardClick);
+
+//   const filmDetailsPopupComponent = new FilmDetailsPopupComponent(film);
+//   filmDetailsPopupComponent.setPopupCloseButtonClickHandler(onCloseButtonClick);
+
+//   render(filmListElement, filmComponent, RenderPosition.BEFOREEND);
+// };
+
+// const renderFilmsLists = (filmsComponent, films) => {
+
+//   const filmsContainer = siteMain.querySelector(`.films`);
+
+//   if (films.length === 0) {
+//     filmsContainer.removeChild(filmsContainer.querySelector(`.films-list`));
+//     render(filmsContainer, new NoFilmsComponent(), RenderPosition.BEFOREEND);
+//     return;
+//   }
+//   render(filmsComponent.getElement(), new FilmListComponent(), RenderPosition.BEFOREEND);
+
+//   const filmListElement = siteMain.querySelector(`.films-list`);
+//   render(filmListElement, new FilmCardContainerComponent(), RenderPosition.BEFOREEND);
+//   const mainFilmsListContainer = filmListElement.querySelector(`.films-list__container`);
+
+//   let showingFilmsCount = SHOWING_CARDS_COUNT_ON_START;
+
+//   films
+//     .slice(0, showingFilmsCount)
+//     .forEach((film) => {
+//       renderFilmCard(mainFilmsListContainer, film);
+//     });
+
+
+//   const showMoreButtonComponent = new ShowMoreBtnComponent();
+
+//   render(filmListElement, showMoreButtonComponent, `beforeend`);
+
+//   showMoreButtonComponent.setClickHandler(() => {
+//     const prevFilmCount = showingFilmsCount;
+//     showingFilmsCount = showingFilmsCount + SHOWING_CARDS_COUNT_BY_BUTTON;
+
+//     films.slice(prevFilmCount, showingFilmsCount)
+//       .forEach((film) => renderFilmCard(mainFilmsListContainer, film));
+//     if (showingFilmsCount >= films.length) {
+//       remove(showMoreButtonComponent);
+//     }
+//   });
+// };
+
+// const filmsComponent = new FilmCardContainerComponent();
+
+// renderFilmsLists(filmsComponent, filmCards);
+
+console.log(``);
 
 // const renderExtraFilms = () => {
